Touch up initial parsing implementation.
